omap4/5: Add support for booting with CH.
    Configuration header(CH) is 512 byte header attached to an OMAP
    boot image that will help ROM code to initialize clocks, SDRAM
    etc and copy U-Boot directly into SDRAM. CH can help us in
    by-passing SPL and directly boot U-boot, hence it's an alternative
    for SPL. However, we intend to support both CH and SPL for OMAP4/5.

    Initialization done through CH is limited and is not equivalent
    to that done by SPL. So U-Boot has to distinguish between the
    two cases and handle them accordingly. This patch takes care
    of doing this.

+++ arch/arm/include/asm/omap_common.h
@@ -80,7 +80,7 @@ struct spl_image_info {
 
 extern struct spl_image_info spl_image;
 
-u32 omap_boot_device(void);
+u8 omap_boot_device(void);
 u32 omap_boot_mode(void);
 
 /* SPL common function s*/
